refactor(ReservationStorage): 예약 정보를 저장하는 Storage 생성
- 그에 따른 Controller 및 테스트 수정
